refactor: AI 분석 비동기 처리 로직의 트랜잭션 전파 문제 해결
- 프록시 내부 호출 문제를 해결하기 위해 로직 분리
